Pass --filter-files to newer isort versions
In isort 4.3.11, the default "skip" behavior changed. Filenames passed
on the command line were no longer subject to "skip" rules from the
configuration file.

isort 4.3.18 provides the `--filter-files` command line option that
restores that behavior. Pass it when we're running at least that version
so that our linter respects the configuration settings.