Add a Python version check to the black linter
When configured, this verifies that we're running the expected Python
version. We check this using the `python3` executable (because black
only support Python 3).

If we don't find the expected Python version, we print a warning and
skip linting with black. This isn't a fatal error for the lint workflow,
unlike most of our other version checks.