Always pass --quiet and --check to black (#31)
This linter is currently written to only check the format of Python
files but not to also autoformat them (which would conflict with the
isort linter; more work here is needed). Given that, we should always
pass these flags to black.

Also, check for the 0 error code, which indicates that black didn't find
any formatting issues in the linted file.