Add NodeExternalLinter base class
`NodeExternalLinter` is an abstract base class for Node-based external linters. It abstracts shared logic to find the node-based binary to be used, whether it is known to `yarn`, present at the expected path in `node_modules`, or installed globally and accessible in `$PATH`.

The existing `ESLint`, `Prettier`, and `PrettierESLint` linters now use this base class.