Handle too much output from graphql-schema-linter, other json errors
When there are a lot of lint errors raised by `graphql-schema-linter`, the output JSON it produces can easily exceed the 2^16 character limit on the `$stdout` contents provided by Arcanist. The resulting truncated string will fail to JSON parse and the script will raise an error.

To correct for this, we check if there were any errors encountered in JSON decoding and raise these explicitly so the user can take action.

For this specific case, we see if the stdout buffer is at its maximum length and, if so, advise the author to run the linter outside of Arcanist and correct issues before running again.