Move `install-instructions` to base `PinterestExternalLinter` class
In 53370a3e2944bd699272fd78438157a1f3c64f34 we introduced a common `install-instructions` configuration option for both Node- and Python-based linters.

In this commit, we move this to a `PinterestExternalLinter` base class so the logic is not duplicated. Aside from the cleanup, this also allows us to specify default behavior for `getUpgradeInstructions()` so that it takes advantage of the custom install instructions, if available.

As it is important to have a version string without any whitespace (which will cause false negatives on compare), we also clean up the version checking functions so that they consistently produce version strings without whitespace using regex.