Resolve the full binary path
When the binary is assumed to be in the $PATH (via resolveBinaryPath()),
getDefaultBinary() will only return the unqualified binary name. Because
our dependency-searching logic here requires the full installation path,
we use Filesystem::resolveBinary() to expand it.

This is perhaps something that NodeExternalLinter::resolveBinaryPath()
should do itself, but I didn't want to change the behavior of any of our
existing linters (just in case).

This change improves the reliability of running `arc lint` outside of
the repository root (or wherever the node packages are installed
relative to).

Also, pull the $bin_root assignment out of the inner loop.