Make Pyright configuration file paths absolute
To resolve issues with configuration files placed in non-standard locations outside the working tree, make these paths absolute relative to the arcanist project root.

This was the implicit behavior previously, so this should not be a breaking change.

Also change the format of command flags from `'--project /path/to/foo.json'` to `'--project=/path/to/foo.json'` to resolve issues with pyright's interpretation of them.